fix(runtime): keep AgentBox observability dependency-safe
