+++ src/shared/dispatch-observability.ts
@@ -1,5 +1,3 @@
-import { wrapRpcError } from "../lib/error-envelope.js";
-
 const MAX_LOG_MESSAGE_LENGTH = 512;
 
 export function compactDispatchLogMessage(value: unknown): string {
@@ -8,19 +6,4 @@ export function compactDispatchLogMessage(value: unknown): string {
   return compact.length <= MAX_LOG_MESSAGE_LENGTH
     ? compact
     : `${compact.slice(0, MAX_LOG_MESSAGE_LENGTH)}…`;
-}
-
-export function summarizeDispatchError(err: unknown): {
-  code: string;
-  status?: number;
-  retriable: boolean;
-  message: string;
-} {
-  const detail = wrapRpcError(err);
-  return {
-    code: detail.code,
-    ...(detail.status != null ? { status: detail.status } : {}),
-    retriable: detail.retriable,
-    message: compactDispatchLogMessage(detail.message),
-  };
 }
